the product details getting fetched from the database(production) to the front end

+++ pagination-web-app/src/main/java/com/lti/training/maven/model/Product.java
@@ -3,7 +3,7 @@ package com.lti.training.maven.model;
 public class Product {
 	private int id;
 	private String name;
-	private double price;
+	private int price;
 	private int quantity;
 	public Product() {
 		
@@ -25,10 +25,10 @@ public class Product {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public double getPrice() {
+	public int getPrice() {
 		return price;
 	}
-	public void setPrice(double price) {
+	public void setPrice(int price) {
 		this.price = price;
 	}
 	public int getQuantity() {
